Fixed major planet culling error and started Item drops

+++ src/nitrogene/util/ZoomEnum.java
@@ -2,12 +2,12 @@ package nitrogene.util;
 
 public enum ZoomEnum {
 	NORMAL (1,1),
-	LARGE (0.8, 1.25),
-	MAP (0.5, 2);
+	LARGE (0.8f, 1.25f),
+	MAP (0.5f, 2f);
 	
-	public double scale;
-	public double inverse;
-	ZoomEnum(double scale, double inverse){
+	public float scale;
+	public float inverse;
+	ZoomEnum(float scale, float inverse){
 		this.scale = scale;
 		this.inverse = inverse;
 	}
